Remove HOC connect function
